+++ CHANGELOG.md
@@ -24,9 +24,7 @@ The types of changes are:
 - Backend System-level Cookie Support [#4383](https://github.com/ethyca/fides/pull/4383)
 - High Level Tracking of Compass System Sync [#4397](https://github.com/ethyca/fides/pull/4397)
 - Erasure support for Qualtrics [#4371](https://github.com/ethyca/fides/pull/4371)
-
-### Added
-- Added Ada Chatbot Connector Erasure [#4382](https://github.com/ethyca/fides/pull/4382)
+- Erasure support for Ada Chatbot [#4382](https://github.com/ethyca/fides/pull/4382)
 
 ### Changed
 - Add filtering and pagination to bulk vendor add table [#4351](https://github.com/ethyca/fides/pull/4351)